requirements.txt
fastapi
uvicorn
groq
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fastmcp
elevenlabs
